OpenCPU wrapper: import config from CSV and convert to JSON

OpenCPU wrapper: import config from CSV and convert to JSON

Usage

opencpu_config_read(
  confDir = system.file("extdata/configs/dummy", package = "kwb.qmra")
)

Arguments

confDir

directory to read configuration files (Default: system.file('extdata/configs/dummy', package = 'kwb.qmra')")

Value

stores configuration in JSON format

Examples

### Example json config file
config_json <- kwb.qmra::opencpu_config_read()
#> Rows: 3 Columns: 10
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): name, type
#> dbl (5): value, min, max, mode, mean
#> lgl (3): sd, meanlog, sdlog
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 37 Columns: 13
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (3): PathogenName, PathogenGroup, type
#> dbl (4): PathogenID, simulate, min, max
#> lgl (6): value, mode, mean, sd, meanlog, sdlog
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 38 Columns: 13
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (4): TreatmentName, TreatmentGroup, PathogenGroup, type
#> dbl (3): TreatmentID, min, max
#> lgl (6): value, mode, mean, sd, meanlog, sdlog
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 5 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (2): TreatmentSchemeName, TreatmentName
#> dbl (2): TreatmentSchemeID, TreatmentID
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 37 Columns: 13
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (9): PathogenName, PathogenGroup, Best fit model*, Host type, Dose units...
#> dbl (4): PathogenID, k, alpha, N50
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> Rows: 37 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): PathogenName
#> dbl (3): PathogenID, infection_to_illness, dalys_per_case
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
head(config_json)
#> [1] "{\n  \"exposure\": [\n    {\n      \"name\": \"number_of_repeatings\",\n      \"type\": \"value\",\n      \"value\": 10\n    },\n    {\n      \"name\": \"number_of_exposures\",\n      \"type\": \"value\",\n      \"value\": 365\n    },\n    {\n      \"name\": \"volume_perEvent\",\n      \"type\": \"triangle\",\n      \"min\": 0.5,\n      \"max\": 3,\n      \"mode\": 1.5,\n      \"mean\": 2\n    }\n  ],\n  \"inflow\": [\n    {\n      \"PathogenID\": 1,\n      \"PathogenName\": \"Bacillus anthracis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 2,\n      \"PathogenName\": \"Burkholderia pseudomallei\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 3,\n      \"PathogenName\": \"Campylobacter jejuni and Campylobacter coli\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 1,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 4,\n      \"PathogenName\": \"Coxiella burnetii\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 5,\n      \"PathogenName\": \"Escherichia coli enterohemorrhagic (EHEC)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 6,\n      \"PathogenName\": \"Escherichia coli\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 7,\n      \"PathogenName\": \"Francisella tularensis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 8,\n      \"PathogenName\": \"Legionella pneumophila\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 9,\n      \"PathogenName\": \"Listeria monocytogenes (Death as response)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 10,\n      \"PathogenName\": \"Listeria monocytogenes (Infection)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 11,\n      \"PathogenName\": \"Listeria monocytogenes (Stillbirths)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 12,\n      \"PathogenName\": \"Mycobacterium avium\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 13,\n      \"PathogenName\": \"Pseudomonas aeruginosa (Contact lens)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 14,\n      \"PathogenName\": \"Pseudomonas aeruginosa (bacterimia)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 15,\n      \"PathogenName\": \"Rickettsia rickettsi\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 16,\n      \"PathogenName\": \"Salmonella Typhi\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 17,\n      \"PathogenName\": \"Salmonella anatum\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 18,\n      \"PathogenName\": \"Salmonella meleagridis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 19,\n      \"PathogenName\": \"Salmonella nontyphoid\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 20,\n      \"PathogenName\": \"Salmonella serotype newport\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 21,\n      \"PathogenName\": \"Shigella\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 22,\n      \"PathogenName\": \"Staphylococcus aureus\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 23,\n      \"PathogenName\": \"Vibrio cholerae\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 24,\n      \"PathogenName\": \"Yersinia pestis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 25,\n      \"PathogenName\": \"Adenovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 26,\n      \"PathogenName\": \"Echovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 27,\n      \"PathogenName\": \"Enteroviruses\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 28,\n      \"PathogenName\": \"Influenza\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 29,\n      \"PathogenName\": \"Lassa virus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 30,\n      \"PathogenName\": \"Poliovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 31,\n      \"PathogenName\": \"Rhinovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 32,\n      \"PathogenName\": \"Rotavirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 1,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 33,\n      \"PathogenName\": \"SARS\",\n      \"PathogenGroup\": \"Viruses\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 34,\n      \"PathogenName\": \"Cryptosporidium parvum and Cryptosporidium hominis\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 35,\n      \"PathogenName\": \"Endamoeba coli\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 36,\n      \"PathogenName\": \"Giardia duodenalis\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"simulate\": 1,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    },\n    {\n      \"PathogenID\": 37,\n      \"PathogenName\": \"Naegleria fowleri\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"simulate\": 0,\n      \"type\": \"uniform\",\n      \"min\": 10,\n      \"max\": 10000\n    }\n  ],\n  \"treatment\": {\n    \"processes\": [\n      {\n        \"TreatmentID\": 1,\n        \"TreatmentName\": \"Conventional clarification\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 0.2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 1,\n        \"TreatmentName\": \"Conventional clarification\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 1,\n        \"TreatmentName\": \"Conventional clarification\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 0.1,\n        \"max\": 3.4\n      },\n      {\n        \"TreatmentID\": 2,\n        \"TreatmentName\": \"Dissolved air flotation\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 0.6,\n        \"max\": 2.6\n      },\n      {\n        \"TreatmentID\": 3,\n        \"TreatmentName\": \"High-rate clarification\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2.8\n      },\n      {\n        \"TreatmentID\": 4,\n        \"TreatmentName\": \"Lime softening\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 4\n      },\n      {\n        \"TreatmentID\": 4,\n        \"TreatmentName\": \"Lime softening\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 0,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 4,\n        \"TreatmentName\": \"Lime softening\",\n        \"TreatmentGroup\": \"Coagulation, flocculation and sedimentation\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 4\n      },\n      {\n        \"TreatmentID\": 5,\n        \"TreatmentName\": \"Granular high-rate filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 0.2,\n        \"max\": 4.4\n      },\n      {\n        \"TreatmentID\": 5,\n        \"TreatmentName\": \"Granular high-rate filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 0.4,\n        \"max\": 3.3\n      },\n      {\n        \"TreatmentID\": 5,\n        \"TreatmentName\": \"Granular high-rate filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 0,\n        \"max\": 3.5\n      },\n      {\n        \"TreatmentID\": 6,\n        \"TreatmentName\": \"Membrane filtration (micro, ultra-, nanofiltration, reverse osmosis)\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 7\n      },\n      {\n        \"TreatmentID\": 6,\n        \"TreatmentName\": \"Membrane filtration (micro, ultra-, nanofiltration, reverse osmosis)\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 2.3,\n        \"max\": 7\n      },\n      {\n        \"TreatmentID\": 6,\n        \"TreatmentName\": \"Membrane filtration (micro, ultra-, nanofiltration, reverse osmosis)\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 6.5\n      },\n      {\n        \"TreatmentID\": 7,\n        \"TreatmentName\": \"Precoat filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 0.2,\n        \"max\": 2.3\n      },\n      {\n        \"TreatmentID\": 7,\n        \"TreatmentName\": \"Precoat filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 3,\n        \"max\": 6.7\n      },\n      {\n        \"TreatmentID\": 7,\n        \"TreatmentName\": \"Precoat filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 1.7\n      },\n      {\n        \"TreatmentID\": 8,\n        \"TreatmentName\": \"Slow sand filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 6\n      },\n      {\n        \"TreatmentID\": 8,\n        \"TreatmentName\": \"Slow sand filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 0.3,\n        \"max\": 5\n      },\n      {\n        \"TreatmentID\": 8,\n        \"TreatmentName\": \"Slow sand filtration\",\n        \"TreatmentGroup\": \"Filtration\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 0.25,\n        \"max\": 4\n      },\n      {\n        \"TreatmentID\": 9,\n        \"TreatmentName\": \"Bank filtration\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 6\n      },\n      {\n        \"TreatmentID\": 9,\n        \"TreatmentName\": \"Bank filtration\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 1,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 9,\n        \"TreatmentName\": \"Bank filtration\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 2.1,\n        \"max\": 8.3\n      },\n      {\n        \"TreatmentID\": 10,\n        \"TreatmentName\": \"Roughing filters\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 0.2,\n        \"max\": 2.3\n      },\n      {\n        \"TreatmentID\": 11,\n        \"TreatmentName\": \"Storage reservoirs\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 0.7,\n        \"max\": 2.2\n      },\n      {\n        \"TreatmentID\": 11,\n        \"TreatmentName\": \"Storage reservoirs\",\n        \"TreatmentGroup\": \"Pretreatment\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 1.4,\n        \"max\": 2.3\n      },\n      {\n        \"TreatmentID\": 12,\n        \"TreatmentName\": \"Chlorine\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 12,\n        \"TreatmentName\": \"Chlorine\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 12,\n        \"TreatmentName\": \"Chlorine\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 13,\n        \"TreatmentName\": \"Chlorine dioxide\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 13,\n        \"TreatmentName\": \"Chlorine dioxide\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 13,\n        \"TreatmentName\": \"Chlorine dioxide\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 14,\n        \"TreatmentName\": \"Ozone\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 14,\n        \"TreatmentName\": \"Ozone\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 14,\n        \"TreatmentName\": \"Ozone\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 2,\n        \"max\": 2\n      },\n      {\n        \"TreatmentID\": 15,\n        \"TreatmentName\": \"UV\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Bacteria\",\n        \"type\": \"uniform\",\n        \"min\": 4,\n        \"max\": 4\n      },\n      {\n        \"TreatmentID\": 15,\n        \"TreatmentName\": \"UV\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Protozoa\",\n        \"type\": \"uniform\",\n        \"min\": 4,\n        \"max\": 4\n      },\n      {\n        \"TreatmentID\": 15,\n        \"TreatmentName\": \"UV\",\n        \"TreatmentGroup\": \"Primary disinfection\",\n        \"PathogenGroup\": \"Viruses\",\n        \"type\": \"uniform\",\n        \"min\": 4,\n        \"max\": 4\n      }\n    ],\n    \"schemes\": [\n      {\n        \"TreatmentSchemeID\": 1,\n        \"TreatmentSchemeName\": \"Berlin (BF + Slow sand)\",\n        \"TreatmentID\": 8,\n        \"TreatmentName\": \"Slow sand filtration\"\n      },\n      {\n        \"TreatmentSchemeID\": 1,\n        \"TreatmentSchemeName\": \"Berlin (BF + Slow sand)\",\n        \"TreatmentID\": 9,\n        \"TreatmentName\": \"Bank filtration\"\n      },\n      {\n        \"TreatmentSchemeID\": 2,\n        \"TreatmentSchemeName\": \"Depth & surface filtration\",\n        \"TreatmentID\": 1,\n        \"TreatmentName\": \"Conventional clarification\"\n      },\n      {\n        \"TreatmentSchemeID\": 2,\n        \"TreatmentSchemeName\": \"Depth & surface filtration\",\n        \"TreatmentID\": 5,\n        \"TreatmentName\": \"Granular high-rate filtration\"\n      },\n      {\n        \"TreatmentSchemeID\": 2,\n        \"TreatmentSchemeName\": \"Depth & surface filtration\",\n        \"TreatmentID\": 15,\n        \"TreatmentName\": \"UV\"\n      }\n    ]\n  },\n  \"doseresponse\": [\n    {\n      \"PathogenID\": 1,\n      \"PathogenName\": \"Bacillus anthracis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0,\n      \"Host type\": \"guinea pig\",\n      \"Dose units\": \"spores\",\n      \"Route\": \"inhalation\",\n      \"Response\": \"death\",\n      \"Reference\": \"Druett 1953\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Bacillus_anthracis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 2,\n      \"PathogenName\": \"Burkholderia pseudomallei\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.328,\n      \"N50\": 5430,\n      \"Host type\": \"C57BL/6 mice and diabetic rat\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"intranasal,intraperitoneal\",\n      \"Response\": \"death\",\n      \"Reference\": \"Liu, Koo et al. 2002 and Brett and Woods 1996\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Burkholderia_pseudomallei:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 3,\n      \"PathogenName\": \"Campylobacter jejuni and Campylobacter coli\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.144,\n      \"N50\": 890,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral (in milk)\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Black et al 1988\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Campylobacter_jejuni_and_Campylobacter_coli:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 4,\n      \"PathogenName\": \"Coxiella burnetii\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.357,\n      \"N50\": 493000000,\n      \"Host type\": \"C57BL/1OScN mice\",\n      \"Dose units\": \"PFU\",\n      \"Route\": \"intraperitoneal\",\n      \"Response\": \"death\",\n      \"Reference\": \"Williams et al, 1982\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Coxiella_burnetii:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 5,\n      \"PathogenName\": \"Escherichia coli enterohemorrhagic (EHEC)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0002,\n      \"Host type\": \"pig\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral (in food)\",\n      \"Response\": \"shedding in feces\",\n      \"Reference\": \"Cornick & Helgerson (2004)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Escherichia_coli_enterohemorrhagic_(EHEC):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 6,\n      \"PathogenName\": \"Escherichia coli\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.155,\n      \"N50\": 2110000,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral (in milk)\",\n      \"Response\": \"positive stool isolation\",\n      \"Reference\": \"DuPont et al. (1971)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Escherichia_coli:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 7,\n      \"PathogenName\": \"Francisella tularensis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0473,\n      \"Host type\": \"monkeys\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"inhalation\",\n      \"Response\": \"death\",\n      \"Reference\": \"Day and Berendt, 1972\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Francisella_tularensis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 8,\n      \"PathogenName\": \"Legionella pneumophila\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0599,\n      \"Host type\": \"guinea pig\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"inhalation\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Muller et al. (1983)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Legionella_pneumophila:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 9,\n      \"PathogenName\": \"Listeria monocytogenes (Death as response)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0,\n      \"Host type\": \"C57B1/6J mice\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"death\",\n      \"Reference\": \"Golnazarian, Donnelly et al. 1989\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Listeria_monocytogenes_(Death_as_response):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 10,\n      \"PathogenName\": \"Listeria monocytogenes (Infection)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.253,\n      \"N50\": 277,\n      \"Host type\": \"C57Bl/6J mice\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Golnazarian\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Listeria_monocytogenes_(Infection):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 11,\n      \"PathogenName\": \"Listeria monocytogenes (Stillbirths)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.0422,\n      \"N50\": 1780000000,\n      \"Host type\": \"pooled\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"stillbirths\",\n      \"Reference\": \"Smith, Williams2007\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Listeria_monocytogenes_(Stillbirths):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 12,\n      \"PathogenName\": \"Mycobacterium avium\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0007,\n      \"Host type\": \"deer\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"O'Brien et al(1976)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Mycobacterium_avium:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 13,\n      \"PathogenName\": \"Pseudomonas aeruginosa (Contact lens)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.19,\n      \"N50\": 18500,\n      \"Host type\": \"white rabbit\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"contact lens\",\n      \"Response\": \"corneal ulceration\",\n      \"Reference\": \"Lawin-Brussel et al. (1993)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Pseudomonas_aeruginosa_(Contact_lens):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 14,\n      \"PathogenName\": \"Pseudomonas aeruginosa (bacterimia)\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0001,\n      \"Host type\": \"Swiss webster mice (5day old)\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"injected in eyelids\",\n      \"Response\": \"death\",\n      \"Reference\": \"Hazlett, Rosen et al. 1978\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Pseudomonas_aeruginosa_(bacterimia):_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 15,\n      \"PathogenName\": \"Rickettsia rickettsi\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.777,\n      \"N50\": 21.3,\n      \"Host type\": \"Pooled data\",\n      \"Dose units\": \"CFU\",\n      \"Response\": \"morbidity\",\n      \"Reference\": \"Saslaw and Carlisle 1966 and Dupont, Hornick et al. 1973\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Rickettsia_rickettsi:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 16,\n      \"PathogenName\": \"Salmonella Typhi\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.175,\n      \"N50\": 1110000,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral, in milk\",\n      \"Response\": \"disease\",\n      \"Reference\": \"Hornick et al. (1966),Hornick et al. (1970)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Salmonella_Typhi:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 17,\n      \"PathogenName\": \"Salmonella anatum\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.318,\n      \"N50\": 37100,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral, with eggnog\",\n      \"Response\": \"positive stool culture\",\n      \"Reference\": \"McCullough and Elsele,1951\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Salmonella_anatum:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 18,\n      \"PathogenName\": \"Salmonella meleagridis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.389,\n      \"N50\": 16800,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral, with eggnog\",\n      \"Response\": \"infection\",\n      \"Reference\": \"McCullough and Eisele 1951,2\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Salmonella_meleagridis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 19,\n      \"PathogenName\": \"Salmonella nontyphoid\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.21,\n      \"N50\": 49.8,\n      \"Host type\": \"mice\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"intraperitoneal\",\n      \"Response\": \"death\",\n      \"Reference\": \"Meynell and Meynell,1958\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Salmonella_nontyphoid:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 20,\n      \"PathogenName\": \"Salmonella serotype newport\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 3.97e-06,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"McCullough and Elsele,1951\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Salmonella_serotype_newport:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 21,\n      \"PathogenName\": \"Shigella\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.265,\n      \"N50\": 1480,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral (in milk)\",\n      \"Response\": \"positive stool isolation\",\n      \"Reference\": \"DuPont et al. (1972b)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Shigella:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 22,\n      \"PathogenName\": \"Staphylococcus aureus\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 7.64e-08,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU/cm2\",\n      \"Route\": \"subcutaneous\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Rose and Haas 1999\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Staphylococcus_aureus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 23,\n      \"PathogenName\": \"Vibrio cholerae\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.25,\n      \"N50\": 243,\n      \"Host type\": \"human\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"oral (with NaHCO3)\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Hornick et al., (1971)\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Vibrio_cholerae:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 24,\n      \"PathogenName\": \"Yersinia pestis\",\n      \"PathogenGroup\": \"Bacteria\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0016,\n      \"Host type\": \"mice\",\n      \"Dose units\": \"CFU\",\n      \"Route\": \"intranasal\",\n      \"Response\": \"death\",\n      \"Reference\": \"Lathem et al. 2005\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Yersinia_pestis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 25,\n      \"PathogenName\": \"Adenovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.607,\n      \"Host type\": \"human\",\n      \"Dose units\": \"TCID50\",\n      \"Route\": \"inhalation\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Couch, Cate et al. 1966\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Adenovirus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 26,\n      \"PathogenName\": \"Echovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 1.06,\n      \"N50\": 922,\n      \"Host type\": \"human\",\n      \"Dose units\": \"PFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Schiff et al.,1984\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Echovirus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 27,\n      \"PathogenName\": \"Enteroviruses\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0037,\n      \"Host type\": \"pig\",\n      \"Dose units\": \"PFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Cliver, 1981\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Enteroviruses:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 28,\n      \"PathogenName\": \"Influenza\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.581,\n      \"N50\": 945000,\n      \"Host type\": \"human\",\n      \"Dose units\": \"TCID50\",\n      \"Route\": \"intranasal\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Murphy et al., 1984 & Murphy et al., 1985\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Influenza:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 29,\n      \"PathogenName\": \"Lassa virus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 2.95,\n      \"Host type\": \"guinea pig\",\n      \"Dose units\": \"PFU\",\n      \"Route\": \"subcutaneous\",\n      \"Response\": \"death\",\n      \"Reference\": \"Jahrling et al., 1982\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Lassa_virus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 30,\n      \"PathogenName\": \"Poliovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.491,\n      \"Host type\": \"human\",\n      \"Dose units\": \"PD50 (mouse paralytic doses)\",\n      \"Route\": \"oral (capsule)\",\n      \"Response\": \"alimentary infection\",\n      \"Reference\": \"Koprowski\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Poliovirus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 31,\n      \"PathogenName\": \"Rhinovirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.221,\n      \"N50\": 1.81,\n      \"Host type\": \"human\",\n      \"Dose units\": \"TCID50 doses\",\n      \"Route\": \"intranasal\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Hendley et al., 1972\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Rhinovirus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 32,\n      \"PathogenName\": \"Rotavirus\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.253,\n      \"N50\": 6.17,\n      \"Host type\": \"human\",\n      \"Dose units\": \"FFU\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Ward et al, 1986\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Rotavirus:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 33,\n      \"PathogenName\": \"SARS\",\n      \"PathogenGroup\": \"Viruses\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0025,\n      \"Host type\": \"mice hACE-2 and A/J\",\n      \"Dose units\": \"PFU\",\n      \"Route\": \"intranasal\",\n      \"Response\": \"death\",\n      \"Reference\": \"DeDiego et al., 2008 &  De Albuquerque et al., 2006\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/SARS:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 34,\n      \"PathogenName\": \"Cryptosporidium parvum and Cryptosporidium hominis\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0572,\n      \"Host type\": \"human\",\n      \"Dose units\": \"oocysts\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Messner et al. 2001\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Cryptosporidium_parvum_and_Cryptosporidium_hominis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 35,\n      \"PathogenName\": \"Endamoeba coli\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"Best fit model*\": \"beta-Poisson\",\n      \"alpha\": 0.101,\n      \"N50\": 341,\n      \"Host type\": \"human\",\n      \"Dose units\": \"Cysts\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Rendtorff 1954\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Endamoeba_coli:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 36,\n      \"PathogenName\": \"Giardia duodenalis\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 0.0199,\n      \"Host type\": \"human\",\n      \"Dose units\": \"Cysts\",\n      \"Route\": \"oral\",\n      \"Response\": \"infection\",\n      \"Reference\": \"Rendtorff 1954\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Giardia_duodenalis:_Dose_Response_Models\"\n    },\n    {\n      \"PathogenID\": 37,\n      \"PathogenName\": \"Naegleria fowleri\",\n      \"PathogenGroup\": \"Protozoa\",\n      \"Best fit model*\": \"exponential\",\n      \"k\": 3.42e-07,\n      \"Host type\": \"mice\",\n      \"Dose units\": \"no of trophozoites\",\n      \"Route\": \"intravenous\",\n      \"Response\": \"death\",\n      \"Reference\": \"Adams et al. 1976 & Haggerty and John 1978\",\n      \"Link\": \"http://qmrawiki.canr.msu.edu/index.php/Naegleria_fowleri:_Dose_Response_Models\"\n    }\n  ],\n  \"health\": [\n    {\n      \"PathogenID\": 1,\n      \"PathogenName\": \"Bacillus anthracis\"\n    },\n    {\n      \"PathogenID\": 2,\n      \"PathogenName\": \"Burkholderia pseudomallei\"\n    },\n    {\n      \"PathogenID\": 3,\n      \"PathogenName\": \"Campylobacter jejuni and Campylobacter coli\",\n      \"infection_to_illness\": 0.7,\n      \"dalys_per_case\": 0.0046\n    },\n    {\n      \"PathogenID\": 4,\n      \"PathogenName\": \"Coxiella burnetii\"\n    },\n    {\n      \"PathogenID\": 5,\n      \"PathogenName\": \"Escherichia coli enterohemorrhagic (EHEC)\"\n    },\n    {\n      \"PathogenID\": 6,\n      \"PathogenName\": \"Escherichia coli\"\n    },\n    {\n      \"PathogenID\": 7,\n      \"PathogenName\": \"Francisella tularensis\"\n    },\n    {\n      \"PathogenID\": 8,\n      \"PathogenName\": \"Legionella pneumophila\"\n    },\n    {\n      \"PathogenID\": 9,\n      \"PathogenName\": \"Listeria monocytogenes (Death as response)\"\n    },\n    {\n      \"PathogenID\": 10,\n      \"PathogenName\": \"Listeria monocytogenes (Infection)\"\n    },\n    {\n      \"PathogenID\": 11,\n      \"PathogenName\": \"Listeria monocytogenes (Stillbirths)\"\n    },\n    {\n      \"PathogenID\": 12,\n      \"PathogenName\": \"Mycobacterium avium\"\n    },\n    {\n      \"PathogenID\": 13,\n      \"PathogenName\": \"Pseudomonas aeruginosa (Contact lens)\"\n    },\n    {\n      \"PathogenID\": 14,\n      \"PathogenName\": \"Pseudomonas aeruginosa (bacterimia)\"\n    },\n    {\n      \"PathogenID\": 15,\n      \"PathogenName\": \"Rickettsia rickettsi\"\n    },\n    {\n      \"PathogenID\": 16,\n      \"PathogenName\": \"Salmonella Typhi\"\n    },\n    {\n      \"PathogenID\": 17,\n      \"PathogenName\": \"Salmonella anatum\"\n    },\n    {\n      \"PathogenID\": 18,\n      \"PathogenName\": \"Salmonella meleagridis\"\n    },\n    {\n      \"PathogenID\": 19,\n      \"PathogenName\": \"Salmonella nontyphoid\"\n    },\n    {\n      \"PathogenID\": 20,\n      \"PathogenName\": \"Salmonella serotype newport\"\n    },\n    {\n      \"PathogenID\": 21,\n      \"PathogenName\": \"Shigella\"\n    },\n    {\n      \"PathogenID\": 22,\n      \"PathogenName\": \"Staphylococcus aureus\"\n    },\n    {\n      \"PathogenID\": 23,\n      \"PathogenName\": \"Vibrio cholerae\"\n    },\n    {\n      \"PathogenID\": 24,\n      \"PathogenName\": \"Yersinia pestis\"\n    },\n    {\n      \"PathogenID\": 25,\n      \"PathogenName\": \"Adenovirus\"\n    },\n    {\n      \"PathogenID\": 26,\n      \"PathogenName\": \"Echovirus\"\n    },\n    {\n      \"PathogenID\": 27,\n      \"PathogenName\": \"Enteroviruses\"\n    },\n    {\n      \"PathogenID\": 28,\n      \"PathogenName\": \"Influenza\"\n    },\n    {\n      \"PathogenID\": 29,\n      \"PathogenName\": \"Lassa virus\"\n    },\n    {\n      \"PathogenID\": 30,\n      \"PathogenName\": \"Poliovirus\"\n    },\n    {\n      \"PathogenID\": 31,\n      \"PathogenName\": \"Rhinovirus\"\n    },\n    {\n      \"PathogenID\": 32,\n      \"PathogenName\": \"Rotavirus\",\n      \"infection_to_illness\": 0.03,\n      \"dalys_per_case\": 0.014\n    },\n    {\n      \"PathogenID\": 33,\n      \"PathogenName\": \"SARS\"\n    },\n    {\n      \"PathogenID\": 34,\n      \"PathogenName\": \"Cryptosporidium parvum and Cryptosporidium hominis\"\n    },\n    {\n      \"PathogenID\": 35,\n      \"PathogenName\": \"Endamoeba coli\"\n    },\n    {\n      \"PathogenID\": 36,\n      \"PathogenName\": \"Giardia duodenalis\",\n      \"infection_to_illness\": 0.3,\n      \"dalys_per_case\": 0.0015\n    },\n    {\n      \"PathogenID\": 37,\n      \"PathogenName\": \"Naegleria fowleri\"\n    }\n  ]\n}"