risk_json: example risk_json object returned by kwb.qmra::opencpu_simulate_risk()

JSON risk result object generated with R script (data-raw/risk_json.R)

data(risk_json)

Format

A json list with all risk results generated by opencpu_simulate_risk, using the dataset `config_json`

Examples

data("risk_json")
risk <- jsonlite::fromJSON(risk_json)
### only show "stats" elements (skip "events" and "total" due to much data)
risk_stats <- list(stats_total = risk$stats_total,
                   stats_logremoval = risk$stats_logremoval)
jsonlite::toJSON(risk_stats, pretty = TRUE)

#> {
#>   "stats_total": [
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dalys_sum",
#>       "min": 0.0001,
#>       "p05": 0.0001,
#>       "p25": 0.0001,
#>       "mean": 0.0002,
#>       "median": 0.0002,
#>       "p75": 0.0003,
#>       "p95": 0.0004,
#>       "max": 0.0004
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dose_sum",
#>       "min": 1,
#>       "p05": 1.45,
#>       "p25": 2.25,
#>       "mean": 3.8,
#>       "median": 3.5,
#>       "p75": 4.75,
#>       "p95": 7,
#>       "max": 7
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "exposure_sum",
#>       "min": 1.9133,
#>       "p05": 2.3836,
#>       "p25": 3.0388,
#>       "mean": 3.716,
#>       "median": 3.7783,
#>       "p75": 4.522,
#>       "p95": 4.9465,
#>       "max": 5.0697
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "illnessProb_sum",
#>       "min": 0.0128,
#>       "p05": 0.0186,
#>       "p25": 0.0287,
#>       "mean": 0.0472,
#>       "median": 0.0435,
#>       "p75": 0.0596,
#>       "p95": 0.0851,
#>       "max": 0.0851
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "infectionProb_sum",
#>       "min": 0.0184,
#>       "p05": 0.0265,
#>       "p25": 0.0408,
#>       "mean": 0.0668,
#>       "median": 0.0617,
#>       "p75": 0.0842,
#>       "p95": 0.1197,
#>       "max": 0.1197
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "inflow_median",
#>       "min": 4872.7872,
#>       "p05": 4883.5106,
#>       "p25": 4938.0158,
#>       "mean": 5090.1941,
#>       "median": 5054.2859,
#>       "p75": 5215.8884,
#>       "p95": 5373.5954,
#>       "max": 5447.6299
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "logreduction_median",
#>       "min": 7.8179,
#>       "p05": 7.8313,
#>       "p25": 7.9004,
#>       "mean": 7.9802,
#>       "median": 7.9863,
#>       "p75": 8.0409,
#>       "p95": 8.1388,
#>       "max": 8.1846
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dalys_sum",
#>       "min": 0.0176,
#>       "p05": 0.0182,
#>       "p25": 0.0192,
#>       "mean": 0.021,
#>       "median": 0.0206,
#>       "p75": 0.0221,
#>       "p95": 0.025,
#>       "max": 0.0266
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dose_sum",
#>       "min": 3754,
#>       "p05": 3958.75,
#>       "p25": 4339.5,
#>       "mean": 5042.2,
#>       "median": 4807.5,
#>       "p75": 5345.25,
#>       "p95": 6745.05,
#>       "max": 7398
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "exposure_sum",
#>       "min": 3867.2196,
#>       "p05": 4013.4301,
#>       "p25": 4330.5065,
#>       "mean": 5070.1129,
#>       "median": 4911.3322,
#>       "p75": 5344.1957,
#>       "p95": 6748.7729,
#>       "max": 7419.3763
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "illnessProb_sum",
#>       "min": 1,
#>       "p05": 1,
#>       "p25": 1,
#>       "mean": 1,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "infectionProb_sum",
#>       "min": 1,
#>       "p05": 1,
#>       "p25": 1,
#>       "mean": 1,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "inflow_median",
#>       "min": 4429.6721,
#>       "p05": 4480.8975,
#>       "p25": 4636.7843,
#>       "mean": 4803.2467,
#>       "median": 4750.7529,
#>       "p75": 4990.4583,
#>       "p95": 5212.5506,
#>       "max": 5231.0166
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "logreduction_median",
#>       "min": 4.0615,
#>       "p05": 4.0617,
#>       "p25": 4.0953,
#>       "mean": 4.1585,
#>       "median": 4.1381,
#>       "p75": 4.2016,
#>       "p95": 4.291,
#>       "max": 4.2969
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dalys_sum",
#>       "min": 0.0026,
#>       "p05": 0.0026,
#>       "p25": 0.0031,
#>       "mean": 0.0036,
#>       "median": 0.0034,
#>       "p75": 0.0042,
#>       "p95": 0.0049,
#>       "max": 0.0052
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dose_sum",
#>       "min": 48,
#>       "p05": 56.55,
#>       "p25": 72.5,
#>       "mean": 88.7,
#>       "median": 85,
#>       "p75": 99.5,
#>       "p95": 132,
#>       "max": 141
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "exposure_sum",
#>       "min": 46.1852,
#>       "p05": 55.4773,
#>       "p25": 77.9714,
#>       "mean": 97.5662,
#>       "median": 96.4384,
#>       "p75": 117.9932,
#>       "p95": 144.1634,
#>       "max": 160.9056
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "illnessProb_sum",
#>       "min": 0.1697,
#>       "p05": 0.1723,
#>       "p25": 0.1977,
#>       "mean": 0.2273,
#>       "median": 0.2155,
#>       "p75": 0.2579,
#>       "p95": 0.2952,
#>       "max": 0.3102
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "infectionProb_sum",
#>       "min": 0.9997,
#>       "p05": 0.9997,
#>       "p25": 0.9999,
#>       "mean": 0.9999,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "inflow_median",
#>       "min": 4783.5281,
#>       "p05": 4786.7481,
#>       "p25": 4842.5935,
#>       "mean": 5002.9931,
#>       "median": 5060.5483,
#>       "p75": 5103.957,
#>       "p95": 5199.2321,
#>       "max": 5231.2913
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "logreduction_median",
#>       "min": 7.0084,
#>       "p05": 7.1066,
#>       "p25": 7.2645,
#>       "mean": 7.3197,
#>       "median": 7.3236,
#>       "p75": 7.4238,
#>       "p95": 7.4936,
#>       "max": 7.5191
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dalys_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0.0001,
#>       "mean": 0.0002,
#>       "median": 0.0002,
#>       "p75": 0.0002,
#>       "p95": 0.0003,
#>       "max": 0.0003
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dose_sum",
#>       "min": 0,
#>       "p05": 0.45,
#>       "p25": 2,
#>       "mean": 2.7,
#>       "median": 3,
#>       "p75": 3,
#>       "p95": 5.1,
#>       "max": 6
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "exposure_sum",
#>       "min": 2.111,
#>       "p05": 2.1791,
#>       "p25": 2.3513,
#>       "mean": 2.7297,
#>       "median": 2.5437,
#>       "p75": 2.7261,
#>       "p95": 3.9823,
#>       "max": 4.9276
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "illnessProb_sum",
#>       "min": 0,
#>       "p05": 0.0058,
#>       "p25": 0.0255,
#>       "mean": 0.0338,
#>       "median": 0.038,
#>       "p75": 0.038,
#>       "p95": 0.0623,
#>       "max": 0.0732
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "infectionProb_sum",
#>       "min": 0,
#>       "p05": 0.0083,
#>       "p25": 0.0364,
#>       "mean": 0.048,
#>       "median": 0.0541,
#>       "p75": 0.0541,
#>       "p95": 0.0881,
#>       "max": 0.1033
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "inflow_median",
#>       "min": 4872.7872,
#>       "p05": 4883.5106,
#>       "p25": 4938.0158,
#>       "mean": 5090.1941,
#>       "median": 5054.2859,
#>       "p75": 5215.8884,
#>       "p95": 5373.5954,
#>       "max": 5447.6299
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "logreduction_median",
#>       "min": 7.2826,
#>       "p05": 7.3059,
#>       "p25": 7.3924,
#>       "mean": 7.4162,
#>       "median": 7.4177,
#>       "p75": 7.4494,
#>       "p95": 7.5126,
#>       "max": 7.5286
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dalys_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 4.4332e-06,
#>       "median": 0,
#>       "p75": 8.8665e-06,
#>       "p95": 0,
#>       "max": 0
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dose_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.5,
#>       "median": 0,
#>       "p75": 1,
#>       "p95": 1.55,
#>       "max": 2
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "exposure_sum",
#>       "min": 0.6102,
#>       "p05": 0.633,
#>       "p25": 0.6789,
#>       "mean": 0.6968,
#>       "median": 0.6977,
#>       "p75": 0.7289,
#>       "p95": 0.7465,
#>       "max": 0.7555
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "illnessProb_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.003,
#>       "median": 0,
#>       "p75": 0.0059,
#>       "p95": 0.0091,
#>       "max": 0.0118
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "infectionProb_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.0098,
#>       "median": 0,
#>       "p75": 0.0197,
#>       "p95": 0.0303,
#>       "max": 0.039
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "inflow_median",
#>       "min": 4429.6721,
#>       "p05": 4480.8975,
#>       "p25": 4636.7843,
#>       "mean": 4803.2467,
#>       "median": 4750.7529,
#>       "p75": 4990.4583,
#>       "p95": 5212.5506,
#>       "max": 5231.0166
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "logreduction_median",
#>       "min": 7.2688,
#>       "p05": 7.2693,
#>       "p25": 7.2805,
#>       "mean": 7.3385,
#>       "median": 7.3315,
#>       "p75": 7.4001,
#>       "p95": 7.413,
#>       "max": 7.4184
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dalys_sum",
#>       "min": 0.0001,
#>       "p05": 0.0002,
#>       "p25": 0.0003,
#>       "mean": 0.0004,
#>       "median": 0.0003,
#>       "p75": 0.0004,
#>       "p95": 0.0007,
#>       "max": 0.0008
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dose_sum",
#>       "min": 1,
#>       "p05": 1.45,
#>       "p25": 3,
#>       "mean": 3.7,
#>       "median": 3,
#>       "p75": 4.75,
#>       "p95": 6.55,
#>       "max": 7
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "exposure_sum",
#>       "min": 2.8927,
#>       "p05": 2.9891,
#>       "p25": 3.3856,
#>       "mean": 4.268,
#>       "median": 4.1712,
#>       "p75": 5.0847,
#>       "p95": 5.6922,
#>       "max": 6.0802
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "illnessProb_sum",
#>       "min": 0.0079,
#>       "p05": 0.0114,
#>       "p25": 0.021,
#>       "mean": 0.0267,
#>       "median": 0.0235,
#>       "p75": 0.0313,
#>       "p95": 0.0484,
#>       "max": 0.054
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "infectionProb_sum",
#>       "min": 0.2634,
#>       "p05": 0.3507,
#>       "p25": 0.574,
#>       "mean": 0.6122,
#>       "median": 0.6003,
#>       "p75": 0.707,
#>       "p95": 0.8499,
#>       "max": 0.8823
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "inflow_median",
#>       "min": 4783.5281,
#>       "p05": 4786.7481,
#>       "p25": 4842.5935,
#>       "mean": 5002.9931,
#>       "median": 5060.5483,
#>       "p75": 5103.957,
#>       "p95": 5199.2321,
#>       "max": 5231.2913
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "logreduction_median",
#>       "min": 7.4494,
#>       "p05": 7.4542,
#>       "p25": 7.4758,
#>       "mean": 7.5383,
#>       "median": 7.554,
#>       "p75": 7.583,
#>       "p95": 7.6188,
#>       "max": 7.6357
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "volume_sum",
#>       "min": 590.4106,
#>       "p05": 591.8517,
#>       "p25": 604.4293,
#>       "mean": 608.406,
#>       "median": 609.1534,
#>       "p75": 614.5369,
#>       "p95": 622.5616,
#>       "max": 622.6079
#>     }
#>   ],
#>   "stats_logremoval": [
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 2.0007,
#>       "p05": 2.2112,
#>       "p25": 3.0135,
#>       "mean": 4.0084,
#>       "median": 3.9861,
#>       "p75": 5.0413,
#>       "p95": 5.7928,
#>       "max": 6
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 0.3098,
#>       "p05": 0.573,
#>       "p25": 1.4986,
#>       "mean": 2.6507,
#>       "median": 2.6374,
#>       "p75": 3.8044,
#>       "p95": 4.7497,
#>       "max": 4.9994
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.2522,
#>       "p05": 0.4368,
#>       "p25": 1.2292,
#>       "mean": 2.1666,
#>       "median": 2.1935,
#>       "p75": 3.102,
#>       "p95": 3.8129,
#>       "max": 3.9897
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 2.0019,
#>       "p05": 2.203,
#>       "p25": 2.9821,
#>       "mean": 3.9931,
#>       "median": 3.9873,
#>       "p75": 5.0019,
#>       "p95": 5.8169,
#>       "max": 5.9983
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 1.0008,
#>       "p05": 1.049,
#>       "p25": 1.2574,
#>       "mean": 1.5001,
#>       "median": 1.5012,
#>       "p75": 1.7515,
#>       "p95": 1.9454,
#>       "max": 1.9999
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 2.1006,
#>       "p05": 2.4125,
#>       "p25": 3.6034,
#>       "mean": 5.1635,
#>       "median": 5.1299,
#>       "p75": 6.7186,
#>       "p95": 8.0033,
#>       "max": 8.2945
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Bacteria",
#>       "min": 0.2001,
#>       "p05": 0.2882,
#>       "p25": 0.6564,
#>       "mean": 1.1014,
#>       "median": 1.0965,
#>       "p75": 1.5552,
#>       "p95": 1.8989,
#>       "max": 1.9998
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Protozoa",
#>       "min": 1.0001,
#>       "p05": 1.0473,
#>       "p25": 1.2416,
#>       "mean": 1.4983,
#>       "median": 1.4972,
#>       "p75": 1.7565,
#>       "p95": 1.9482,
#>       "max": 1.9999
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.101,
#>       "p05": 0.2761,
#>       "p25": 0.9431,
#>       "mean": 1.7638,
#>       "median": 1.7602,
#>       "p75": 2.5889,
#>       "p95": 3.244,
#>       "max": 3.3995
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 0.2005,
#>       "p05": 0.4237,
#>       "p25": 1.2641,
#>       "mean": 2.3179,
#>       "median": 2.3349,
#>       "p75": 3.3603,
#>       "p95": 4.1856,
#>       "max": 4.3999
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 0.4032,
#>       "p05": 0.5377,
#>       "p25": 1.1072,
#>       "mean": 1.8447,
#>       "median": 1.8488,
#>       "p75": 2.5643,
#>       "p95": 3.1507,
#>       "max": 3.2991
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.0029,
#>       "p05": 0.1648,
#>       "p25": 0.8405,
#>       "mean": 1.739,
#>       "median": 1.7222,
#>       "p75": 2.6635,
#>       "p95": 3.3296,
#>       "max": 3.4996
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Bacteria",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Protozoa",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Viruses",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     }
#>   ]
#> }