Simulate: inflow

Simulate: inflow

simulate_inflow(config, debug = TRUE, lean = FALSE)

Arguments

config
as retrieved by config_read()

debug
print debug information (default: TRUE)

lean
if TRUE, only the "events" are returned (in a reduced version, i.e. without column PathogenName and with all ID columns being of class integer), otherwise a list with the events in element events" and the corresponding parameters in element paras. The default is FALSE, i.e. events and parameters are returned in a list.

Value

list with parameters of user defined random distribution and corresponding values