Create random distribution

Create random distribution

create_random_distribution(type = "uniform", number_of_repeatings = 1,
  number_of_events = 365, value = 10, min = 10, max = 1000,
  mean = (min + max)/2, sdev = (max - mean)/qnorm(0.975),
  meanlog = mean(log((min + max)/2)), sdlog = sd(log((c(min, max)))),
  mode = (min + max)/2, debug = TRUE)

Arguments

type
"uniform" calls runif(), "triangle" calls EnvStats::rtri(), "lognorm" calls rlnorm() and "norm" calls rnorm(), (default: "uniform")

number_of_repeatings
how often should the random distribution with the same parameters be generated (default: 1)

number_of_events
number of events

value
constant value (no random number), gets repeated number_of_events times (if 'type' = 'value')

min
minimum value (default: 10), only used if 'type' is "runif" or "triangle"

max
maximum value (default: 1000), only used if 'type' is "runif" or "triangle"

mean
mean value (default: mean of min & max value), only used if 'type' is "norm"

sdev
standard deviation (default: (max-mean) / qnorm(0.975)), only used if 'type' is "norm"

meanlog
log mean value (default: mean of log min & max value), only used if 'type' is "lognorm"

sdlog
standard deviation (default: sd of log min & max value), only used if 'type' is "lognorm"

mode
(default: mean of min & max), only used if 'type' is "triangle"

debug
print debug information (default: TRUE)

Value

list with parameters of user defined random distribution and corresponding values

See also