risk_dummy_json: example risk_json object returned by kwb.qmra::opencpu_simulate_risk()

JSON risk result object generated with R script (data-raw/risk_json.R)

data(risk_dummy_json)

Format

A json list with all risk results generated by opencpu_simulate_risk, using the dataset `config_dummy_json`

Examples

data("risk_dummy_json")
risk <- jsonlite::fromJSON(risk_dummy_json)
### only show "stats" elements (skip "events" and "total" due to much data)
risk_stats <- list(stats_total = risk$stats_total,
                   stats_logremoval = risk$stats_logremoval)
jsonlite::toJSON(risk_stats, pretty = TRUE)

#> {
#>   "stats_total": [
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dalys_sum",
#>       "min": 0.0001,
#>       "p05": 0.0001,
#>       "p25": 0.0002,
#>       "mean": 0.0002,
#>       "median": 0.0003,
#>       "p75": 0.0003,
#>       "p95": 0.0003,
#>       "max": 0.0003
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dose_sum",
#>       "min": 1,
#>       "p05": 1.9,
#>       "p25": 3,
#>       "mean": 4.2,
#>       "median": 4.5,
#>       "p75": 5.75,
#>       "p95": 6,
#>       "max": 6
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "exposure_sum",
#>       "min": 2.4141,
#>       "p05": 2.576,
#>       "p25": 2.8244,
#>       "mean": 3.5872,
#>       "median": 3.2268,
#>       "p75": 4.4081,
#>       "p95": 4.9695,
#>       "max": 5.1303
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "illnessProb_sum",
#>       "min": 0.0128,
#>       "p05": 0.0242,
#>       "p25": 0.038,
#>       "mean": 0.0518,
#>       "median": 0.0551,
#>       "p75": 0.0702,
#>       "p95": 0.0732,
#>       "max": 0.0732
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "infectionProb_sum",
#>       "min": 0.0184,
#>       "p05": 0.0344,
#>       "p25": 0.0541,
#>       "mean": 0.0733,
#>       "median": 0.078,
#>       "p75": 0.0991,
#>       "p95": 0.1033,
#>       "max": 0.1033
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "inflow_median",
#>       "min": 4653.3805,
#>       "p05": 4748.3957,
#>       "p25": 4933.9815,
#>       "mean": 5075.777,
#>       "median": 5116.1414,
#>       "p75": 5196.9233,
#>       "p95": 5369.9929,
#>       "max": 5452.4783
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "logreduction_median",
#>       "min": 7.7191,
#>       "p05": 7.7974,
#>       "p25": 7.9129,
#>       "mean": 7.9632,
#>       "median": 7.9904,
#>       "p75": 8.042,
#>       "p95": 8.058,
#>       "max": 8.0603
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dalys_sum",
#>       "min": 0.0168,
#>       "p05": 0.0173,
#>       "p25": 0.0196,
#>       "mean": 0.021,
#>       "median": 0.0208,
#>       "p75": 0.0229,
#>       "p95": 0.0245,
#>       "max": 0.0246
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dose_sum",
#>       "min": 3423,
#>       "p05": 3757.8,
#>       "p25": 4339.25,
#>       "mean": 5017.6,
#>       "median": 4814.5,
#>       "p75": 5712.25,
#>       "p95": 6475.35,
#>       "max": 6996
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "exposure_sum",
#>       "min": 3505.3026,
#>       "p05": 3798.133,
#>       "p25": 4286.3844,
#>       "mean": 5006.6546,
#>       "median": 4750.8716,
#>       "p75": 5751.648,
#>       "p95": 6441.1899,
#>       "max": 6895.9955
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "illnessProb_sum",
#>       "min": 1,
#>       "p05": 1,
#>       "p25": 1,
#>       "mean": 1,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "infectionProb_sum",
#>       "min": 1,
#>       "p05": 1,
#>       "p25": 1,
#>       "mean": 1,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "inflow_median",
#>       "min": 4710.4905,
#>       "p05": 4739.7344,
#>       "p25": 4791.3663,
#>       "mean": 4964.467,
#>       "median": 4971.1302,
#>       "p75": 5106.3243,
#>       "p95": 5228.0037,
#>       "max": 5228.0943
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "logreduction_median",
#>       "min": 3.8804,
#>       "p05": 3.8953,
#>       "p25": 3.9922,
#>       "mean": 4.1318,
#>       "median": 4.121,
#>       "p75": 4.2702,
#>       "p95": 4.3587,
#>       "max": 4.3695
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dalys_sum",
#>       "min": 0.0035,
#>       "p05": 0.0036,
#>       "p25": 0.0038,
#>       "mean": 0.0042,
#>       "median": 0.0041,
#>       "p75": 0.0044,
#>       "p95": 0.0051,
#>       "max": 0.0054
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dose_sum",
#>       "min": 55,
#>       "p05": 59.5,
#>       "p25": 95.5,
#>       "mean": 109.1,
#>       "median": 99,
#>       "p75": 117.25,
#>       "p95": 180.4,
#>       "max": 202
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "exposure_sum",
#>       "min": 51.5657,
#>       "p05": 52.5866,
#>       "p25": 91.004,
#>       "mean": 112.7844,
#>       "median": 104.0966,
#>       "p75": 122.1383,
#>       "p95": 196.8526,
#>       "max": 202.18
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "illnessProb_sum",
#>       "min": 0.2198,
#>       "p05": 0.2269,
#>       "p25": 0.2411,
#>       "mean": 0.2603,
#>       "median": 0.2544,
#>       "p75": 0.2725,
#>       "p95": 0.3067,
#>       "max": 0.3235
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "infectionProb_sum",
#>       "min": 1,
#>       "p05": 1,
#>       "p25": 1,
#>       "mean": 1,
#>       "median": 1,
#>       "p75": 1,
#>       "p95": 1,
#>       "max": 1
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "inflow_median",
#>       "min": 4462.2161,
#>       "p05": 4548.8245,
#>       "p25": 4727.0798,
#>       "mean": 4914.3861,
#>       "median": 4923.0337,
#>       "p75": 5037.1534,
#>       "p95": 5308.4335,
#>       "max": 5461.4185
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "logreduction_median",
#>       "min": 7.0088,
#>       "p05": 7.127,
#>       "p25": 7.2791,
#>       "mean": 7.3232,
#>       "median": 7.3125,
#>       "p75": 7.4015,
#>       "p95": 7.5006,
#>       "max": 7.5273
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dalys_sum",
#>       "min": 0.0001,
#>       "p05": 0.0001,
#>       "p25": 0.0001,
#>       "mean": 0.0002,
#>       "median": 0.0002,
#>       "p75": 0.0003,
#>       "p95": 0.0004,
#>       "max": 0.0004
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "dose_sum",
#>       "min": 1,
#>       "p05": 1.45,
#>       "p25": 2,
#>       "mean": 3.7,
#>       "median": 3.5,
#>       "p75": 5.75,
#>       "p95": 6,
#>       "max": 6
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "exposure_sum",
#>       "min": 1.6829,
#>       "p05": 1.8597,
#>       "p25": 2.5111,
#>       "mean": 2.8584,
#>       "median": 2.6394,
#>       "p75": 3.137,
#>       "p95": 4.1788,
#>       "max": 4.7658
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "illnessProb_sum",
#>       "min": 0.0128,
#>       "p05": 0.0186,
#>       "p25": 0.0255,
#>       "mean": 0.0462,
#>       "median": 0.0442,
#>       "p75": 0.0706,
#>       "p95": 0.074,
#>       "max": 0.0746
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "infectionProb_sum",
#>       "min": 0.0184,
#>       "p05": 0.0265,
#>       "p25": 0.0364,
#>       "mean": 0.0653,
#>       "median": 0.0627,
#>       "p75": 0.0996,
#>       "p95": 0.1043,
#>       "max": 0.1052
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "inflow_median",
#>       "min": 4653.3805,
#>       "p05": 4748.3957,
#>       "p25": 4933.9815,
#>       "mean": 5075.777,
#>       "median": 5116.1414,
#>       "p75": 5196.9233,
#>       "p95": 5369.9929,
#>       "max": 5452.4783
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "logreduction_median",
#>       "min": 7.1716,
#>       "p05": 7.1908,
#>       "p25": 7.266,
#>       "mean": 7.391,
#>       "median": 7.3989,
#>       "p75": 7.5208,
#>       "p95": 7.6096,
#>       "max": 7.6542
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Bacteria",
#>       "PathogenID": 3,
#>       "PathogenName": "Campylobacter jejuni and Campylobacter coli",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dalys_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 7.0932e-06,
#>       "median": 0,
#>       "p75": 0,
#>       "p95": 0,
#>       "max": 0
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "dose_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.8,
#>       "median": 0,
#>       "p75": 1.5,
#>       "p95": 3.1,
#>       "max": 4
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "exposure_sum",
#>       "min": 0.607,
#>       "p05": 0.6075,
#>       "p25": 0.6292,
#>       "mean": 0.7022,
#>       "median": 0.6966,
#>       "p75": 0.7415,
#>       "p95": 0.8261,
#>       "max": 0.8306
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "illnessProb_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.0047,
#>       "median": 0,
#>       "p75": 0.0088,
#>       "p95": 0.0182,
#>       "max": 0.0234
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "infectionProb_sum",
#>       "min": 0,
#>       "p05": 0,
#>       "p25": 0,
#>       "mean": 0.0155,
#>       "median": 0,
#>       "p75": 0.0293,
#>       "p95": 0.0596,
#>       "max": 0.0765
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "inflow_median",
#>       "min": 4710.4905,
#>       "p05": 4739.7344,
#>       "p25": 4791.3663,
#>       "mean": 4964.467,
#>       "median": 4971.1302,
#>       "p75": 5106.3243,
#>       "p95": 5228.0037,
#>       "max": 5228.0943
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "logreduction_median",
#>       "min": 7.1917,
#>       "p05": 7.2191,
#>       "p25": 7.3043,
#>       "mean": 7.3374,
#>       "median": 7.3217,
#>       "p75": 7.3775,
#>       "p95": 7.4715,
#>       "max": 7.542
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Protozoa",
#>       "PathogenID": 36,
#>       "PathogenName": "Giardia duodenalis",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dalys_sum",
#>       "min": 0,
#>       "p05": 0.0001,
#>       "p25": 0.0002,
#>       "mean": 0.0003,
#>       "median": 0.0003,
#>       "p75": 0.0005,
#>       "p95": 0.0006,
#>       "max": 0.0006
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "dose_sum",
#>       "min": 0,
#>       "p05": 0.9,
#>       "p25": 2,
#>       "mean": 3.3,
#>       "median": 3,
#>       "p75": 5,
#>       "p95": 5.55,
#>       "max": 6
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "events",
#>       "min": 365,
#>       "p05": 365,
#>       "p25": 365,
#>       "mean": 365,
#>       "median": 365,
#>       "p75": 365,
#>       "p95": 365,
#>       "max": 365
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "exposure_sum",
#>       "min": 2.3873,
#>       "p05": 2.8178,
#>       "p25": 3.5178,
#>       "mean": 4.0366,
#>       "median": 3.9838,
#>       "p75": 4.62,
#>       "p95": 5.2298,
#>       "max": 5.2663
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "illnessProb_sum",
#>       "min": 0,
#>       "p05": 0.0071,
#>       "p25": 0.0157,
#>       "mean": 0.0243,
#>       "median": 0.0235,
#>       "p75": 0.0339,
#>       "p95": 0.0404,
#>       "max": 0.0416
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "infectionProb_sum",
#>       "min": 0,
#>       "p05": 0.2058,
#>       "p25": 0.4574,
#>       "mean": 0.5614,
#>       "median": 0.5815,
#>       "p75": 0.7426,
#>       "p95": 0.7981,
#>       "max": 0.8104
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "inflow_median",
#>       "min": 4462.2161,
#>       "p05": 4548.8245,
#>       "p25": 4727.0798,
#>       "mean": 4914.3861,
#>       "median": 4923.0337,
#>       "p75": 5037.1534,
#>       "p95": 5308.4335,
#>       "max": 5461.4185
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "logreduction_median",
#>       "min": 7.3243,
#>       "p05": 7.3481,
#>       "p25": 7.4304,
#>       "mean": 7.4605,
#>       "median": 7.4664,
#>       "p75": 7.5158,
#>       "p95": 7.5335,
#>       "max": 7.5363
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "PathogenGroup": "Viruses",
#>       "PathogenID": 32,
#>       "PathogenName": "Rotavirus",
#>       "key": "volume_sum",
#>       "min": 596.2623,
#>       "p05": 597.6677,
#>       "p25": 606.7883,
#>       "mean": 610.3644,
#>       "median": 608.1938,
#>       "p75": 613.861,
#>       "p95": 626.7628,
#>       "max": 630.3435
#>     }
#>   ],
#>   "stats_logremoval": [
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 2.0014,
#>       "p05": 2.1995,
#>       "p25": 2.9496,
#>       "mean": 3.9522,
#>       "median": 3.9282,
#>       "p75": 4.9495,
#>       "p95": 5.8008,
#>       "max": 5.9951
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 0.3012,
#>       "p05": 0.5401,
#>       "p25": 1.4754,
#>       "mean": 2.6434,
#>       "median": 2.6309,
#>       "p75": 3.8296,
#>       "p95": 4.7445,
#>       "max": 4.9998
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 8,
#>       "TreatmentName": "Slow sand filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.2506,
#>       "p05": 0.422,
#>       "p25": 1.147,
#>       "mean": 2.1011,
#>       "median": 2.0889,
#>       "p75": 3.0362,
#>       "p95": 3.7995,
#>       "max": 3.9984
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 2.0045,
#>       "p05": 2.2171,
#>       "p25": 3.0059,
#>       "mean": 4.0067,
#>       "median": 3.9859,
#>       "p75": 5.0289,
#>       "p95": 5.7863,
#>       "max": 5.9987
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 1,
#>       "p05": 1.0486,
#>       "p25": 1.2518,
#>       "mean": 1.5053,
#>       "median": 1.5114,
#>       "p75": 1.7531,
#>       "p95": 1.949,
#>       "max": 2
#>     },
#>     {
#>       "TreatmentSchemeID": 1,
#>       "TreatmentSchemeName": "Berlin (BF + Slow sand)",
#>       "TreatmentID": 9,
#>       "TreatmentName": "Bank filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 2.1003,
#>       "p05": 2.4273,
#>       "p25": 3.6641,
#>       "mean": 5.2238,
#>       "median": 5.2596,
#>       "p75": 6.7806,
#>       "p95": 8.0118,
#>       "max": 8.2997
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Bacteria",
#>       "min": 0.2003,
#>       "p05": 0.2924,
#>       "p25": 0.6525,
#>       "mean": 1.1022,
#>       "median": 1.0981,
#>       "p75": 1.5512,
#>       "p95": 1.9016,
#>       "max": 1.9996
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Protozoa",
#>       "min": 1.0005,
#>       "p05": 1.052,
#>       "p25": 1.2548,
#>       "mean": 1.4979,
#>       "median": 1.4931,
#>       "p75": 1.7456,
#>       "p95": 1.9424,
#>       "max": 2
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 1,
#>       "TreatmentName": "Conventional clarification",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.1006,
#>       "p05": 0.2664,
#>       "p25": 0.8919,
#>       "mean": 1.7376,
#>       "median": 1.7231,
#>       "p75": 2.5986,
#>       "p95": 3.2464,
#>       "max": 3.3992
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Bacteria",
#>       "min": 0.2034,
#>       "p05": 0.4124,
#>       "p25": 1.2579,
#>       "mean": 2.3108,
#>       "median": 2.3515,
#>       "p75": 3.3656,
#>       "p95": 4.2145,
#>       "max": 4.3988
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Protozoa",
#>       "min": 0.4004,
#>       "p05": 0.5533,
#>       "p25": 1.1291,
#>       "mean": 1.8465,
#>       "median": 1.8459,
#>       "p75": 2.5724,
#>       "p95": 3.1581,
#>       "max": 3.2993
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 5,
#>       "TreatmentName": "Granular high-rate filtration",
#>       "PathogenGroup": "Viruses",
#>       "min": 0.0009,
#>       "p05": 0.1595,
#>       "p25": 0.8356,
#>       "mean": 1.7211,
#>       "median": 1.7191,
#>       "p75": 2.5967,
#>       "p95": 3.3046,
#>       "max": 3.4973
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Bacteria",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Protozoa",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     },
#>     {
#>       "TreatmentSchemeID": 2,
#>       "TreatmentSchemeName": "Depth & surface filtration",
#>       "TreatmentID": 15,
#>       "TreatmentName": "UV",
#>       "PathogenGroup": "Viruses",
#>       "min": 4,
#>       "p05": 4,
#>       "p25": 4,
#>       "mean": 4,
#>       "median": 4,
#>       "p75": 4,
#>       "p95": 4,
#>       "max": 4
#>     }
#>   ]
#> }